feat: allow for textobject queries to work again + add cpp queries
This reverts setting max_start_depth to 1, this previously used to make
matches not work as expected. This is because it would go to the parent
node and the parent node contains the previous function as a `field`.
This matches the field and returns without matching the actual function.
(Matching for example the function just before the function call that is
selected). Now we match based on if we are in range as well (if there
exists an identifying capture). If there is no identifying capture, it
will do the same behaviour (wrong matching), however, this is better
than not matching at all.

a